Use AuthTokenSection in App
Using type declarations for Nextcloud global "OC" for now.

The eslint rule set @nextcloud/eslint-config/typescript was added to
prevent:

   ERROR in [eslint]
   /app/src/components/AuthTokenSection.vue
     35:27  error  "@nextcloud/initial-state" is not published  n/no-unpublished-import
     36:32  error  "@nextcloud/l10n" is not published           n/no-unpublished-import

I can only assume that that rule is disabled in the rule set we extend
from.

The error actually complains about the module not being part of
"dependencies" in package.json, which is the case, however.

The internet calls for allowing such modules, which IMO defeats the
purpose.

The rule is also present in Nextcloud's core config.